[#200] equal jitter for `AdvanceRetryPolicy`
